requirements.txt
Flask==2.0.1
python-docx==0.8.11
PyPDF2==1.26.0
python-dotenv==0.19.0